# src/api.yml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Usage api
  description: Usage api service APIs
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /metering
paths:
...
  /v1/report/iasp/{reportUrn}:
    post:
      description: |
        upload report for iasp
      operationId: postReportIASP
      x-eov-operation-handler: iasp-report
      parameters:
        - in: query
          name: accountId
          schema:
            type: string
          required: true
          description: account associated with the report
        - in: query
          name: startDate
          schema:
            type: number
          required: false
          description: Start date for the report (ms since epoch). If provided, endDate is mandatory
        - in: query
          name: endDate
          schema:
            type: number
          required: false
          description: End date for the report (ms since epoch). If provided, startDate is mandatory
        - in: path
          name: reportUrn
          schema:
            type: integer
          required: true
          description: unique identifer for this report. To amend an existing report, submit a new report with the same reportUrn.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: report accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accepted'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      description: |
        download report for iasp
      operationId: getIaspReport
      x-eov-operation-handler: iasp-report
      parameters:
        - in: query
          name: accountId
          schema:
            type: string
          description: account associated with the report. Must be passed for reportUrn value other than 'template'.
        - in: path
          name: reportUrn
          schema:
            type: string
          required: true
          description: unique identifer for the report
      responses:
        '200':
          description: 'report file'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v2/replay:
    post:
      description: |
        Replay metrics - reprocess a set of usage files as if they were just uploaded. The files to be replayed are identified by an array of requestId's and eventId's. Only superusers can call this API.
      operationId: handleReplayArray
      x-eov-operation-handler: replay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 1000
              uniqueItems: true
              items:
                type: object
                properties:
                  requestId:
                    description: Unique ID for a request. Identifies the request to be replayed.
                    type: string
                    pattern: '^[\da-f]{64}$'
                  eventId:
                    description: ID for an event. If provided, only replays this event from the request.
                    type: string
                  accountOrPrefix:
                    description: Specifies the Tenant containing the request. If this field is not specified, the default Tenant will be used.
                    type: string
                    nullable: true
                required:
                  - requestId
      responses:
        '202':
          description: all accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplayAccepted'
        '207':
          description: partially accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplayAccepted'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v2/replay/batch:
    post:
      description: Replay batch - reprocess a set of usage events received during given timestamp. If limit is specified, it limits the number of usage events. Only superusers can call this API.
      operationId: handleReplayBatch
      x-eov-operation-handler: replay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchReplay'
      responses:
        '202':
          description: all success
          content:
            application/json:
              schema:
                type: object
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Accepted:
      required:
        - requestId
      properties:
        requestId:
          type: string
...
    ReplayAccepted:
      description: Array of responses, one for each requestId/eventId, indicating whether it was successfully started or failed. If the overall status code is 207, then only some replays were successfully started.
      type: array
      items:
        type: object
        properties:
          requestId:
            description: Unique ID for the request.
            type: string
            pattern: '^[\da-f]{64}$'
          eventId:
            description: Event ID replayed.
            type: string
          accountOrPrefix:
            description: Identifies the Tenant containing the request.
            type: string
            nullable: true
          httpCode:
            description: True if this request will be replayed. False if for any reason we could not replay.
            type: number
        required:
          - requestId
          - httpCode
    Error:
      required:
        - errorCode
        - message
      properties:
        errorCode:
          oneOf:
            - type: string
            - type: number
        message:
          oneOf:
            - type: string
            - type: number
        details:
          type: object
...
    BatchReplay:
      required:
        - start
      properties:
        start:
          type: number
        end:
          type: number
        accountId:
          type: string
        accountOrPrefix:
          type: string
        isAborted:
          type: boolean
        limit:
          type: number
        requestReplays:
          type: boolean
          default: false
      additionalProperties: false
...